refact: 주식 시세(실시간 체결가) 조회를 위한 웹소켓 세션 연결/해제 스케줄러 리팩토링
 - 패키지 변경: application.websocket -> application.stockquote
 - 클래스명 변경: KisWebSocketConnectionScheduler -> StockQuoteConnectionScheduler
   - application 레이어 내 외부 인프라에 종속적인 네이밍을 제거하고, 도메인 중심적 이름으로 변경
 - 한국투자증권 국내휴장일조회 REST API를 통한 개장 여부 기반 세션 연결 로직 추가
